+++ models/event.js
@@ -22,7 +22,7 @@ var FieldSchema = new Schema({
   id: { type: Number, required: true },
   archived: { type: Boolean, required: false},
   title: { type: String, required: true },
-  type: { type: String, required: true, enum: ['textfield', 'numberfield', 'email','password','radio','dropdown','userDropdown','date','geometry','textarea','checkbox','hidden'] },
+  type: { type: String, required: true, enum: ['textfield', 'numberfield', 'email', 'password', 'radio', 'dropdown', 'multiselectdropdown', 'date', 'geometry', 'textarea', 'checkbox', 'hidden'] },
   value: { type: Schema.Types.Mixed, required: false },
   name: { type: String, required: true },
   required: { type: Boolean, required: true },
@@ -287,7 +287,7 @@ exports.getById = function(id, options, callback) {
         });
       });
     }
-    
+
     async.series(filters, function(err) {
       if (err) return callback(err);