+++ public/app/admin/events/event.edit.form.controller.js
@@ -45,20 +45,53 @@ function AdminEventEditFormController($rootScope, $scope, $location, $filter, $r
 
   $scope.fileUploadOptions = {};
 
-  $scope.fieldTypes = {
-    textfield : 'Textfield',
-    numberfield: 'Numberfield',
-    email : 'E-mail',
-    password : 'Password',
-    radio : 'Radio Buttons',
-    dropdown : 'Dropdown',
-    userDropdown: 'User Dropdown',
-    date : 'Date',
-    geometry : 'Geometry',
-    textarea : 'Text Area',
-    checkbox : 'Checkbox',
-    hidden : 'Hidden'
-  };
+  $scope.fieldTypes = [{
+    name: 'checkbox',
+    title: 'Checkbox'
+  },{
+    name: 'date',
+    title: 'Date'
+  },{
+    name: 'email',
+    title: 'Email'
+  },{
+    name: 'hidden',
+    title: 'Hidden'
+  },{
+    name: 'geometry',
+    title: 'Location'
+  },{
+    name: 'numberfield',
+    title: 'Number'
+  },{
+    name: 'password',
+    title: 'Password'
+  },{
+    name: 'radio',
+    title: 'Radio Buttons'
+  },{
+    name: 'dropdown',
+    title: 'Select'
+  },{
+    name: 'multiselectdropdown',
+    title: 'Multiple Select',
+    hidden: true
+  },{
+    name: 'textfield',
+    title: 'Text'
+  },{
+    name: 'textarea',
+    title: 'Text Area'
+  },{
+    name: 'userDropdown',
+    title: 'User Select'
+  },{
+    name: 'multiSelectUserDropdown',
+    title: 'User Multiple Select',
+    hidden: true
+  }];
+
+  var fieldNameMap = _.indexBy($scope.fieldTypes, 'name');
 
   $scope.newField = newField();
 
@@ -81,11 +114,11 @@ function AdminEventEditFormController($rootScope, $scope, $location, $filter, $r
   };
 
   $scope.getTypeValue = function(field) {
-    if (field.type === 'dropdown' && $scope.isMemberField(field)) {
-      return $scope.fieldTypes.userDropdown;
+    if ($scope.isMemberField(field)) {
+      return field.type === 'dropdown' ? fieldNameMap.userDropdown : fieldNameMap.multiSelectUserDropdown;
     }
 
-    return $scope.fieldTypes[field.type];
+    return fieldNameMap[field.type].title;
   };
 
   // create new field button click
@@ -99,6 +132,10 @@ function AdminEventEditFormController($rootScope, $scope, $location, $filter, $r
     if ($scope.newField.type === 'userDropdown') {
       $scope.event.form.userFields.push($scope.newField.name);
       $scope.newField.type = 'dropdown';
+    }
+
+    if ($scope.newField.type === 'dropdown' && $scope.newField.multiselect) {
+      $scope.newField.type = 'multiselectdropdown';
     }
 
     $scope.onRequiredChanged($scope.newField);
@@ -340,7 +377,7 @@ function AdminEventEditFormController($rootScope, $scope, $location, $filter, $r
 
   // decides whether field options block will be shown (true for dropdown and radio fields)
   $scope.showAddOptions = function(field) {
-    return field.type === "radio" || field.type === "dropdown";
+    return field.type === 'radio' || field.type === 'dropdown'  || field.type === 'multiselectdropdown';
   };
 
   $scope.hideAddOptions = function(field) {