+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ##### Bug Fixes
 * Added form import error handling on web client.
+* Fixed feed layout height in Firefox.
 
 ## [4.2.1](https://github.com/ngageoint/mage-server/releases/tag/v4.2.1) (04-22-2016)
 ##### Features